tagutil: centralize tagutil test
